+++ lib/src/adapters/rewarded_interstitial_ad_adapter.dart
@@ -1,3 +1,4 @@
+import 'package:flutter/material.dart';
 import 'package:google_mobile_ads/google_mobile_ads.dart';
 import 'package:google_mobile_ads_adapter/src/adapters/dismissible_ad_adapter.dart';
 
@@ -6,10 +7,16 @@ class RewardedInterstitialAdAdapter
   RewardedInterstitialAdAdapter(
     super.id, {
     super.request,
+    super.onAdInitialized,
+    super.onAdDismissed,
+    super.onAdClicked,
+    super.onAdShowedFullScreenContent,
+    super.onAdImpression,
+    super.onAdFailedToShowFullScreenContent,
   });
 
   @override
-  Future<void> getAd() => RewardedInterstitialAd.load(
+  Future<void> getAd(BuildContext context) => RewardedInterstitialAd.load(
         adUnitId: id,
         request: request,
         rewardedInterstitialAdLoadCallback: RewardedInterstitialAdLoadCallback(
@@ -18,7 +25,7 @@ class RewardedInterstitialAdAdapter
             ad.fullScreenContentCallback = fullScreenContentCallback;
             return onAdLoaded.call(ad);
           },
-          onAdFailedToLoad: onAdFailedToLoad,
+          onAdFailedToLoad: (error) => onAdFailedToLoad.call(context, error),
         ),
       );
 }